Clean up the project
Remove files that are not needed (anymore), e.g. backup files that stem 
from the era before this project was version controlled with Git.

Additionally remove files that are not used and that need to be 
completely revised, but where it is foreseeable that this revision will 
not be done in the near future.