Lower minimum required Sphinx version
Lower the minimum required Sphinx version from 4.0 to 3.0 again.

Note that Sphinx version 4.1.0 does not render the html logo and favicon
due to a bug (https://github.com/sphinx-doc/sphinx/issues/9438).
However, this version is not explicitly excluded from the requirements
file, because pip tries to avoid to install this version automatically.