[structure.py]: Add Weighted Center Function
Add a function `wcenter` to calculate the weighted center of an
MDAnalysis AtomGroup using its center method.  This function is very
similar to `mdtools.structure.cog` and `mdtools.structure.com`, which
calculate the center of geometry and mass, respectively, of an
MDAnalysis AtomGroup using its center_of_geometry and center_of_mass
methods.