[dtrj.py]: `trans_per_state`: Refactor and New Features
Refactor the code of `mdtools.dtrj.trans_per_state` to be similar to the
code of `mdtools.dtrj.trans_per_state_vs_time`.

Allow to discriminate state transitions where the difference between the
final and the initial state is equal to a given integer.

Allow to respect the minimum image convention when determining the
direction of state transitions.