[numpy_helper_functions.py]: New Function `take`
Add a new function `mdtools.numpy_helper_functions.take` that takes a slice
from a given array along a given axis.  This function is similar to function
`numpy.take`.  The difference is that this function only supports slicing and
not fancy indexing but therefore always returns a view instead of a copy.