[plot_ener_dist.py]: Don't Warn About Non-Gaussian Distributions
Do *not* warn the user anymore if a selected energy value has a
non-Gaussian distribution, because (depending on the simulated ensemble)
the distributions don't have to be Gaussian.  For instance, the
temperature in an NVT ensemble has a skewness of sqrt(8/3N) and a
kurtosis of 3*(1+4/3N), where N is the number of particles.  For small
numbers of particles this differs from zero what we would expect for a
Gaussian distribution.

Reference: Holian, Voter, Ravelo, Thermostatted molecular dynamics: How
to avoid the Toda demon hidden in Nose-Hoover dynamics, Physical Review
E, 1995, 52, 3, 2338-2347.