[attribute_hist.py]: Fix Bug When Using Custom Bins
When using custom bins, not all samples must lie within the bin edges,
but still an error message was raised when this was happening.  This
commit fixes the wrong error message.

Additionally, for the histogram of the Euclidean norms it is ensured
that only positive bin edges are used.

The comments in the output files about the binning is made more
informative.