Use src Layout
Change the project structure from the flat layout to the src layout.

Move the MDTools Matplotlib plotting style into the package so that it
will be contained in the binary distribution.  This should prevent the
bug that the plotting style cannot be find by MDTools after
installation.

For a comparison of the flat and src layout see.
https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/