[plot.py]: Only Rely on `importlib_resources` Backport
`importlib.resources` requires an `__init__.py` file to be present in
each data directory (https://bugs.python.org/issue34534,
https://gitlab.com/python-devs/importlib_resources/-/issues/58).  This
is undesirable, because this exposes the subdirectory as exportable
package (https://stackoverflow.com/a/58941536).  The backport
`importlib_resources` does not require an `__init__.py`.  Because of
this and because MDTools should support Python versions 3.6 to 3.9, we
will only rely on the backport `importlib_resources` even if MDTools is
installed in Python version 3.9.