[.flake8]: Ignore Error Code B028
Ignore error code B028 that was introduced/changed in `flake8-bugbear`
version `23.2.13`.

B028: No explicit stacklevel keyword argument found.  The warn method
from the warnings module uses a stacklevel of 1 by default.  This will
only show a stack trace for the line on which the warn method is called.
It is therefore recommended to use a stacklevel of 2 or greater to
provide more information to the user.