[attribute_hist.py]: Fix Calculation of the Average Mass of the Selected Compounds
Fix the calculation of the average mass of the selected compounds.
Before, the mass was the average mass of all selected atoms.  Now, the
mass is the average mass of all selected compounds (as stated in the
documentation).

This is only relevant for calculating the temperature of the selected
compounds from the Maxwell-Boltzmann fit of the speed distribution.