[run_time_info.py]: Refactor Function `dtrj_trans_info`
Refactor the function `dtrj_trans_info`.  Now the input array is not
transposed, which should be more memory-friendly, because a copy of the
input array is avoided if possible.  However, this might come at the
cost of computational speed, because the memory layout of the array
might not be beneficial for the computations.  Because the input arrays
can become very large, optimizing memory consumption is here more
important than optimizing speed.