[state_lifetime.py]: New Option `--intermittency`
Add a new command-line option `--intermittency` that allows to specify a
maximum number of frames a compound is allowed to leave its state whilst
still being considered to be in this state provided that it returns to
this state after the given number of frames.  In other words, a compound
is only considered to have left its state if it has left it for at least
the given number of frames.