[generate_dtrj.py]: Change Lifetime Distribution
Instead of sampling the lifetimes from a gamma distribution, sample the
lifetimes from the generalized gamma distribution.

Add a new command-line option `--discard` that allows the user to
discard a given number of frames at the beginning of the trajectory to
emulate a left-censoring of the state lifetimes or a discarded
equilibration period.