[dtrj.py]: Function `trans_rate_per_state`: Skip Doctest
This is just a workaround!!

Skip three doctests of the function `mdtools.dtrj.trans_rate_per_state`
when running on a Linux system with Python version 3.9, because these
examples *sometimes* fail during the GitHub CI workflow for this
specific OS-Python combination.  This seems somehow related to caching
but couldn't be resolved so far.